pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "askdev"
version = "1.0.0"
description = "Android Error CLI Assistant: Explain and fix Android errors from your terminal."
authors = [
    { name = "askdev contributors", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]",
    "python-dotenv",
    "requests",
    "rich"
]
classifiers=[
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Bug Tracking",
    "Topic :: Utilities"
]
keywords=["android", "error", "cli", "ai", "fix", "log", "stacktrace"]

[project.urls]
"Homepage" = "https://github.com/divython/askdev-cli"

[project.scripts]
askdev = "askdev.cli:run_main"
ad = "askdev.cli:run_main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = ["pytest"]
